Sprint 396: 48 feature headers, warning fixes, test fixes, dark theme, format detection
- Add 48 new Engine headers across AI/Cache/Core/Decoders/GPU/Memory/Pipeline/Plugin/Utils
- Fix 3 compiler warnings (C4244, C4319, C4100) for zero-warning build
- Fix Test_S395_SystemInfoCollector (populate GPU in snapshot)
- Fix Test_S396_VisualQualityPredictor (dot-prefixed extensions)
- All 2920 tests pass, 0 errors, 0 warnings
- ExplorerLens.py: dark theme, format detection, audio/video decoders
- LENSManager: dark mode helper, RegManager format handlers

+++ Engine/Core/GlobalShortcutManager.h
@@ -57,7 +57,7 @@ public:
     }
 
     bool HandleHotkey(uint32_t id) {
-        auto cb = m_callbacks.find(id);
+        auto cb = m_callbacks.find(static_cast<uint8_t>(id));
         if (cb != m_callbacks.end() && cb->second) {
             cb->second();
             return true;
